+++ module-apps/Application.cpp
@@ -25,6 +25,13 @@
 
 #include "common_data/EventStore.hpp"
 
+namespace app::audioConsts
+{
+    const std::vector<audio::PlaybackType> typesToMute = {audio::PlaybackType::Notifications,
+                                                          audio::PlaybackType::CallRingtone,
+                                                          audio::PlaybackType::TextMessageRingtone};
+} // namespace app::audioConsts
+
 namespace app
 {
 
@@ -507,13 +514,16 @@ namespace app
                                 const audio::Profile::Type &profileType,
                                 const audio::PlaybackType &playbackType)
     {
-        const auto ret = AudioServiceAPI::SetVolume(this, value, profileType, playbackType);
+        const auto ret = AudioServiceAPI::SetSetting(this, audio::Setting::Volume, value, profileType, playbackType);
         return ret == audio::RetCode::Success;
     }
 
     bool Application::adjustCurrentVolume(const int step)
     {
         audio::Volume volume;
+        if (step < 0) {
+            AudioServiceAPI::Stop(this, app::audioConsts::typesToMute);
+        }
         auto ret = AudioServiceAPI::GetSetting(this, audio::Setting::Volume, volume);
         if (ret == audio::RetCode::Success) {
             ret = ((static_cast<int>(volume) + step) < 0)